requirements.txt
# Core Framework
streamlit>=1.28.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Vector Database & Search
qdrant-client
rank-bm25

# LLM & Embeddings
openai>=1.3.0
cohere>=4.37.0
tiktoken>=0.5.0

# Document Processing
pypdf>=3.17.0
langchain-text-splitters>=0.0.1

# Caching
redis>=5.0.0

# Testing & Evaluation (for agents and CI/CD)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
deepeval>=0.20.0

# Additional Utilities
requests>=2.31.0
numpy>=1.24.0
pathlib2;python_version<"3.4"
typing-extensions>=4.8.0
